# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Estudio Guzmán API
  version: 1.0.0
  description: API REST privada. Las mutaciones autenticadas requieren cookie, Origin permitido y x-csrf-token.
servers:
  - url: /api/v1
security:
  - sessionCookie: []
paths:
...
  /auth/login:
    post:
      security: []
      requestBody: { required: true, content: { application/json: { schema: { type: object, required: [email, password], properties: { email: { type: string, format: email }, password: { type: string } } } } } }
      responses: { '200': { description: Sesión iniciada }, '401': { $ref: '#/components/responses/Problem' }, '429': { $ref: '#/components/responses/Problem' } }
  /auth/logout:
    post: { parameters: [ { $ref: '#/components/parameters/Csrf' } ], responses: { '204': { description: Sesión revocada }, '403': { $ref: '#/components/responses/Problem' } } }
  /auth/me:
    get: { responses: { '200': { description: Usuario y permisos efectivos }, '401': { $ref: '#/components/responses/Problem' } } }
...
  /cases:
    get: { responses: { '200': { description: Expedientes paginados } } }
    post: { parameters: [ { $ref: '#/components/parameters/Csrf' } ], responses: { '201': { description: Expediente creado } } }
  /cases/{caseId}:
    parameters: [ { $ref: '#/components/parameters/CaseId' } ]
    get: { responses: { '200': { description: Expediente completo }, '404': { $ref: '#/components/responses/Problem' } } }
    patch: { parameters: [ { $ref: '#/components/parameters/Csrf' } ], responses: { '200': { description: Expediente actualizado }, '409': { $ref: '#/components/responses/Problem' } } }
...
components:
  securitySchemes:
    sessionCookie: { type: apiKey, in: cookie, name: eg_session }
  parameters:
    Csrf: { name: x-csrf-token, in: header, required: true, schema: { type: string } }
    CaseId: { name: caseId, in: path, required: true, schema: { type: string, format: uuid } }
  schemas:
    Problem:
      type: object
      required: [type, title, status, detail, instance, requestId]
      properties: { type: { type: string, format: uri }, title: { type: string }, status: { type: integer }, detail: { type: string }, instance: { type: string }, requestId: { type: string, format: uuid } }
...
  responses:
    Problem: { description: Error RFC 7807, content: { application/problem+json: { schema: { $ref: '#/components/schemas/Problem' } } } }